Added PolicyBottomSheet folder : for restruct
